+++ lib/components/matrix.dart
@@ -24,6 +24,7 @@ import 'avatar.dart';
 
 class Matrix extends StatefulWidget {
   static const String callNamespace = 'chat.fluffy.jitsi_call';
+  static const String defaultHomeserver = 'tchncs.de';
 
   final Widget child;